Move the content of AnalyticsConfig into GoogleAnalytics4Config and UniversalAnalyticsConfig.
